# internal/api/v1/swagger/swagger.yml
basePath: /v1
consumes:
- application/json
host: localhost:8080
info:
  description: things-service service
  version: "1.0"
paths:
  /things:
    get:
      description: Retrieves a thing
      operationId: getThingRQ
      parameters:
      - format: uint64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/getThingRS'
        "400":
          $ref: '#/responses/badRequestRS'
        "500":
          $ref: '#/responses/serverErrorRS'
      summary: Retrieves a thing
      tags:
      - things
    post:
      description: Creates a thing
      operationId: createThing
      parameters:
      - in: body
        name: Body
        schema:
          properties:
            birth_date:
              format: date-time
              type: string
              x-go-name: Birth_date
            email:
              type: string
              x-go-name: Email
            id:
              format: uint64
              type: integer
              x-go-name: ID
            name:
              type: string
              x-go-name: Name
            phone:
              type: string
              x-go-name: Phone
            start_day:
              format: int64
              type: integer
              x-go-name: Start_day
          type: object
      responses:
        "200":
          $ref: '#/responses/CreateThingRS'
        "400":
          $ref: '#/responses/badRequestRS'
        "500":
          $ref: '#/responses/serverErrorRS'
      summary: Creates a thing
      tags:
      - things
produces:
- application/json
responses:
  CreateThingRS:
    description: CreateThingRS represents the response for Swagger documentation
    schema:
      properties:
        birth_date:
          format: date-time
          type: string
          x-go-name: Birth_date
        email:
          type: string
          x-go-name: Email
        id:
          format: uint64
          type: integer
          x-go-name: ID
        name:
          type: string
          x-go-name: Name
        phone:
          type: string
          x-go-name: Phone
        start_day:
          format: int64
          type: integer
          x-go-name: Start_day
      type: object
schemes:
- http
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: api-key
    type: apiKey
swagger: "2.0"